Improved prescan consistency, subframing and made window size computations more robust
x prescan is now always 2 unless the first row of data is not good (3x3 binning with quad readout);
I did this by using a table for the amount to skip. Unfortunately it is not possible
to skip those remaining two rows of prescan (or the one bad column of data in 3x binning).

Size computations are now more robust and self-consistent. In particular 3x3 data size
is the same for 1 amp or 4 amps. This makes things easier to predict, at the expense
of losing one row and column of data (in 3x3 binning with one amp).

Subframing now skips a few extra rows to eliminate some odd junk, making the subframes look better.

Fixed a bug in setFullWindow (window size was not being set in CameraConfig due to a typo)
and fixed Set Full Window in runCamera.py (config.binFac must be set before calling setFullWindow).