Improve FITS keywords
Add _estExpSec to Camera
Fix writing of EXPTIME keyword so 0 is used for bias exposures and estimated time is used if expTime < 0
Add ESTEXPTM keyword, which is only written if expTime argument is not specified to saveImage function.
Write DATASEC[n] and BIASSEC[n] keywords. The [n] is only used for quad readout.